% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/functions.R
\docType{package}
\name{GMTHexBinR}
\alias{GMTHexBinR}
\alias{GMTHexBinR-package}
\title{Functions to plot hexbins in GMT}
\description{
Produce a visual representation of position data with hexbins using the [General Mapping Tool](LINK)
}
\details{
The main function is:
     \describe{
       \item{\code{\link{hexPlot}}:}{to produce a hexmap from a bunch of position data}
     }

And main helper functions are:
     \describe{
       \item{\code{\link{readDouglas}}:}{to read data filtered by the Douglas filter}
     }
}
\author{
Scott Sherrill-Mix, \email{shescott@upenn.edu}
}

