% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/functions.R
\name{hexPlot}
\alias{hexPlot}
\title{Generate hexmaps from position data}
\usage{
hexPlot(daybrs, outFile, hexPerDegree = 1, limits = NULL, extraCmd = NULL,
  landMaskCmd = NULL, hardLimit = NULL, hexMax = NULL,
  logCounts = FALSE, debug = FALSE, maxInterp = 7, gmtDir = "",
  contourFile = NULL, contourDepth = -200, dayScale = 1, interp = 1,
  addNumberToName = TRUE, proportion = FALSE, showMax = TRUE,
  proportionByAnimal = FALSE, seasonLimit = 0, uniqueAnimals = FALSE,
  outlineCount = NULL, propCount = NULL, addPlus = TRUE,
  animalWeights = NULL, ...)
}
\arguments{
\item{daybrs:Data}{frame containing 'animal','rdate','lat','lon' (e.g. the output from readDouglas)}

\item{outFile:}{File location to write postscript map generated by GMT to}

\item{hexPerDegree:}{How many hexs should fit horizontally in 1 degree of longitude at the equator}

\item{limits:}{A vector of (southernMostLatitude,northernMostLatitude,westernMostLongitude,easternMostLongitude) for final hexMap or NULL to select automatically}

\item{extraCmd:}{If not NULL, vector of system commands to run after other GMT commands and before scale (e.g. extra labels with pstext)}

\item{landMaskCmd:}{If not NULL, vector of system commands to run after hex output and before pscoast (e.g. polygons over hexes)}

\item{hardLimit:}{A vector of (southernMostLatitude,northernMostLatitude,westernMostLongitude,easternMostLongitude) for cropping hexs or NULL to not crop}

\item{hexMax:}{Sets hexs with counts > hexMax to hexMax (for creating maps with the highest hex counts being something like "100+")}

\item{logCounts:}{If true, base color scale on logged hex counts}

\item{debug:}{If true, browser() before returning from makeHexs()}

\item{maxInterp:}{Maximum number of missing days to fill between two data points (missing days<=maxInterp)}

\item{gmtDir:}{System directory for calling GMT functions (e.g. if pscoast is in /home/bin/ and /home/bin/ isn't in path then gmtDir="/home/bin/")}

\item{contourFile:}{If not NULL, run pscontour at contourDepth meters based on this GMT elevation file}

\item{contourDepth:}{Depth to draw contour in meters}

\item{dayScale:}{Divide day totals by this number (useful for intervals less than 1 day)}

\item{interp:}{# of positions per day to interpolate}

\item{addNumberToName:}{Add number of days and turtles to file name?}

\item{proportion:}{Divide counts by sum(counts)}

\item{showMax:}{show maximum on scale ticks}

\item{proportionByAnimal:}{weight points to add up to one for each animal}

\item{seaonLimit:}{When using proportionByAnimal, if less than seasonLimit points then total animal weight = # points/seasonLimit otherwise 1}

\item{uniqueAnimals:}{Count unique occurrences of animal in a hex}

\item{outlineCount:}{Draw an outline around all hexes with count >= outlineCount}

\item{propCount:}{Draw an outline around enough hexs to cover propCount proportion of the data (overrides outlineCount)}

\item{addPlus:}{Add a "+" to maximum label in scale (if using hexMax)}

\item{animalWeights:}{Vector (with names of animals) of custom weighting for animals if proportionByAnimal is TRUE (e.g. monthly plots of proportion)}

\item{...:}{Any other arguments to be passed to runGMT}
}
\value{
Vector of (lowXLim,highXlim,lowYLim,highYLim,maxHexCount) can be stored and passed as limits={return}, hardLimit={return} to further hexPlot calls to use the same range of latitude and longitude
}
\description{
Generates various intermediary files (detailed in other functions) for use in plotting
If daybrs contains column 'stateDep' generates animalTable.csv and dayTable.csv showing number of animal and animal-days from each state
If proportionByAnimal writes weights for each animal to 'proportionWeights.csv'
Generates postscript hexmap plot in outFile
}

