% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/functions.R
\name{hsl2rgb}
\alias{hsl2rgb}
\title{Convert hsl to rgb}
\usage{
hsl2rgb(hue, saturation, lightness)
}
\arguments{
\item{hue}{Hue 0-1 in hsl color scale}

\item{saturation}{Saturation 0-1 in hsl color scale}

\item{lightness}{Lightness 0-1 in hsl color svale}
}
\value{
Vector of red green blue in 0-255 scale
}
\description{
Convert hsl color to rgb
}

