% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/functions.R
\name{lagger}
\alias{lagger}
\title{Lag a vector}
\usage{
lagger(data)
}
\arguments{
\item{data}{A vector}
}
\value{
A vector of same length as data with each datum moved one to the right and an NA in position 1
}
\description{
Return a vector lagged one to the right
}

