% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/functions.R
\name{makeHexs}
\alias{makeHexs}
\title{Generate hexs}
\usage{
makeHexs(lat, lon, lonBase = -45, hexPerDegree = 1, border = 5,
  file = "hexs.dat", hardLimit = NULL, hexMax = NULL, logCounts = FALSE,
  debug = FALSE, scale = 1, proportion = FALSE, showMax = FALSE,
  weights = rep(1, length(lat)), uniqueCounter = NULL, addPlus = TRUE)
}
\arguments{
\item{lat}{A vector of latitudes of length N}

\item{lon}{A vector of longitudes of length N}

\item{lonBase}{Longitude for centering calculations}

\item{hexPerDegree}{How many hexs should fit horizontally in 1 degree of longitude at the equator}

\item{border}{How many degrees of space to leave around the hexs (if hardLimit is NULL)}

\item{file}{File to lat/lon of hex positions to}

\item{hardLimit}{A vector of (southernMostLatitude,northernMostLatitude,westernMostLongitude,easternMostLongitude) (or NULL) to filter hexs}

\item{hexMax}{Sets hexs with counts > hexMax to hexMax (for creating maps with the highest hex counts being something like "100+")}

\item{logCounts}{If true base color scale on logged hex counts}

\item{debug}{If true, browser() before returning}

\item{scale}{Divide day totals by this number (useful for intervals less than 1 day)}

\item{proportion}{Divide counts by sum(counts)}

\item{showMax}{show maximum on scale ticks?}

\item{weights}{Weight for each point}

\item{uniqueCounter}{If not NULL, count unique occurrences of uniqueCounter in a hex}

\item{addPlus}{Add a "plus" to maximum label in scale if using hexMax}
}
\value{
Vector for use in further functions of (lowXLim,highXlim,lowYLim,highYLim,maxHexCount)
}
\description{
Writes lat/lon color information to file specified by file parameter in a format ready to use in GMT (but probably easy to parse by other programs too)
Makes test map of hex positions and data in testhex.eps
Makes GMT color table scale.cpt for later use
Calls makeTicks() which will output tick position for scalebar for later use
}

