% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/functions.R
\name{makeScale}
\alias{makeScale}
\title{Make a scale in GMT}
\usage{
makeScale(width, outFile, scaleLabel = "Days Turtles Recorded in Hex",
  margin = 1, scaleHeight = 0.8, xpos = margin/2, ypos = -scaleHeight *
  1.5, gmtDir = "")
}
\arguments{
\item{width}{width of plot area}

\item{outFile}{plot file}

\item{scaleLabel}{title of scale}

\item{margin}{space for margin on either side of scale}

\item{scaleHeight}{height of scale}
}
\value{
Invisible NULL
}
\description{
Makes scale based on ticks.dat and labels.dat
}

