% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/functions.R
\name{makeTicks}
\alias{makeTicks}
\title{Makes ticks for use in GMT scale bar}
\usage{
makeTicks(counts, isLog = FALSE, showMax = TRUE, hexMax = NULL,
  addMin = TRUE, adjustScale = 1, addPlus = TRUE)
}
\arguments{
\item{counts}{Vector of counts to make ticks for}

\item{isLog}{If true counts have previously been transformed round(log(counts)*100)}

\item{showMax}{Add the maximum hex count to the scale and filter out any labels too close to it}

\item{hexMax}{Counts have previously been capped at hexMax. Also adds "+" to {hexMax} label (e.g. "100+") if isLog is false}

\item{addMin}{If true makes sure the minimum scale value is at most 1 (0 if isLog is TRUE)}

\item{addPlus}{Add a "+" to {hexMax} label?}
}
\value{
Invisible NULL
}
\description{
Writes tick positions to ticks.dat for later use in runGMT
Writes labels and label positions to labels.dat for later use in runGMT
}

