% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/functions.R
\name{readDouglas}
\alias{readDouglas}
\title{Read in Douglas filtered files}
\usage{
readDouglas(path = ".", search = "br[0-9][^.]+[0-9]\\\\.txt",
  ignoreErrors = FALSE, extraColumns = NULL)
}
\arguments{
\item{path}{System path to directory containing Douglas outputted files}

\item{search}{Regular expression to select files (defaults to br files)}

\item{ignoreErrors}{If true does not check for correct deployment information (lc94=DP)}

\item{extraColumns}{A vector of extra column names (present in the br files) that should also be returned (should be consistent over a day [e.g. deployday not time])}
}
\value{
Data frame with median position for each day present in br file with columns c('animal', 'ptt', 'date', 'month', 'year', 'rdate', 'season', 'dir','file','deployday', extraColumns)
}
\description{
Sources code in filter.R after reading in br files for custom filtering (see example file)
}

