% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/functions.R
\name{red2bluehsl}
\alias{red2bluehsl}
\title{Generate hsl color red-blue color scale}
\usage{
red2bluehsl(n, min = 1)
}
\arguments{
\item{n}{Maximum count}

\item{min}{Minimum count}
}
\value{
Data frame with rows 1:n and columns 'count','red','green','blue' (rgb columns in 0-255 scale)
}
\description{
Generate color scale for 1:n colors in the hsl color scheme going evenly from hue=.31 to 1 (s=1,v=.5)
}

