% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/functions.R
\name{runGMT}
\alias{runGMT}
\title{Run GMT}
\usage{
runGMT(limits, outFile = "output.ps", width = 6, dataFile = "hexs.dat",
  gmtDir = "", scale = TRUE, contourFile = NULL, contourDepth = -200,
  extraCmd = NULL, landMaskCmd = NULL, landMask = TRUE, ps2eps = FALSE,
  scaleLabel = "Days Turtles Recorded in Hex", contourPen = "1/120ta",
  concatenate = FALSE, xpos = 3, ypos = 1.8,
  posOptions = ifelse(concatenate, "-O", ""), gmtsetOptions = NULL,
  annotateBorders = "WSen", borderColor = 255, pscoastOptions = "")
}
\arguments{
\item{limits}{Vector of (lowXLim,highXlim,lowYLim,highYLim) probably from makeHexs or passed in from hexPlot}

\item{outFile}{File location to output GMT postscript to}

\item{width}{Width of output figure in inches}

\item{dataFile}{File location of hex color and lat/lon data previously generated by makeHexs}

\item{gmtDir}{System directory for calling GMT functions (e.g. if pscoast is in /home/bin/ and /home/bin/ isn't in path then gmtDir="/home/bin/")}

\item{scale}{If TRUE make a scale based on files output by makeTicks (called in makeHexs)}

\item{contourFile}{If not NULL, run pscontour at contourDepth meters based on this GMT elevation file}

\item{contourDepth}{Depth to draw contour in meters}

\item{extraCmd}{If not NULL, vector of system commands to run after other GMT commands and before scale (e.g. extra labels with pstext)}

\item{landMaskCmd}{If not NULL, vector of system commands to run prior to final psCoast (e.g. polygons overlaid by coast)}

\item{landMask}{If TRUE, hexs overlapping land are covered by land else hexs overlap land and only the land border is plotted over them}

\item{ps2eps}{If TRUE, after all other commands calls the program ps2eps to convert .ps to .eps (Make sure you have ps2eps installed and in path if you use this)}

\item{scaleLabel}{Label for scale if scale is TRUE}

\item{contourPen}{A string or vector of strings specifying pen attributes for grd contour to trace the contour file}

\item{concatenate}{Add output to existing file?}

\item{xpos}{Shift plot origin right xpos inches}

\item{ypos}{Shift plot origin up ypos inches}

\item{posOptions}{Options for position (e.g. -O)}

\item{gmtsetOptions}{Options for gmtset}

\item{annotateBorders}{A string like "WSen" specifying which borders to annotate as in psbasemap}

\item{borderColor}{color for border of hexes}

\item{pscoastOptions}{command line options for pscoast}
}
\value{
Invisible NULL
}
\description{
Takes files generated by previous functions and calls mapping program GMT to generate output postscript file
IMPORTANT: You must have GMT installed for this function to work
}

