% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/functions.R
\name{systemOut}
\alias{systemOut}
\title{Sends concatenated command to system}
\usage{
systemOut(..., debug = FALSE)
}
\arguments{
\item{...}{Various parameters to be concatenated into a system command}

\item{debug}{If true prints command to screen (and still calls system)}
}
\value{
Return code returned by OS
}
\description{
Concatenates arguments and sends to system() for executionj
}
\seealso{
\code{\link{paste}}, \code{\link{system}}
}

