% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/functions.R
\name{fill.missing.days}
\alias{fill.missing.days}
\title{Fill in missing days}
\usage{
fill.missing.days(id, date, lat, lon, maxinterp = 7, interp = 1)
}
\arguments{
\item{id}{Vector of length N of animal identifiers (so the last day of one animal is interpolated to the first day of the next)}

\item{date}{Vector of length N of either an rdate or some other integer}

\item{lat}{Vector of length N of latitudes}

\item{lon}{Vector of length N of longitudes}

\item{maxinterp}{Maximum number of missing days to fill between two data points (<=maxinterp)}
}
\value{
Data frame of interpolated date/positions with columns id, data, lat, lon (note: original positions are not returned)
}
\description{
Interpolate between known positions to fill in missing days
}

